ready to test

+++ core/core_model/src/main/java/com/android/core_model/UpdateProduct.kt
@@ -3,7 +3,7 @@ package com.android.core_model
 data class UpdateProduct(
     val id: Int,
     val title: String? = null,
-    val price: Double? = null,
+    val price: Float? = null,
     val description: String? = null,
     val category: String? = null,
     val image: String? = null
